pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-enterprise-starter"
version = "0.1.0"
description = "Enterprise-grade open-source MLOps starter (Colab-safe)"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "mlflow",
  "pyyaml",
  "typer",
  "fastapi",
  "uvicorn",
  "prometheus-client",
  "joblib",
  "requests",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
mlops = "mlops_enterprise.cli:app"
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true